Fix token check bug in binding thread
When we receive the token from the client, we get it without the
0 char at the end. Since we lookup the token in the dns_map as soon as
we receive it and since the dns_map uses strcmp to compare tokens, it
would never return a match.
The solution was to append a 0 char to the received token and use that
token to lookup in the dns_map.